fix: merge mode forwarding into proxy.ts (Next.js 16 compat)
Next.js 16 replaced middleware.ts with proxy.ts — both cannot coexist.
Move the atlas-mode cookie → x-atlas-mode header forwarding into the
existing proxy.ts via withModeHeader() helper. All NextResponse.next()
calls now forward the mode header for server component access.
