+++ packages/web/src/proxy.ts
@@ -1,20 +1,25 @@
 /**
  * Next.js 16 proxy (replaces the middleware.ts convention from Next.js 15).
  *
- * Redirects unauthenticated users to /signup and authenticated users away
- * from auth pages. Only active when NEXT_PUBLIC_ATLAS_AUTH_MODE is "managed"
- * (Better Auth). Other auth modes (none, simple-key, byot) skip the proxy
- * entirely since they don't use Better Auth session cookies.
+ * Two responsibilities:
+ * 1. Auth redirects — redirects unauthenticated users to /signup and
+ *    authenticated users away from auth pages. Only active when
+ *    NEXT_PUBLIC_ATLAS_AUTH_MODE is "managed" (Better Auth).
+ * 2. Mode forwarding — reads the `atlas-mode` cookie and forwards it as
+ *    an `x-atlas-mode` request header so server components can access the
+ *    resolved mode without parsing cookies.
  *
- * This is an optimistic check — it reads the session cookie without hitting
+ * Auth is an optimistic check — it reads the session cookie without hitting
  * the database. Actual auth enforcement happens at the API layer.
  */
 
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 import { getSessionCookie } from "better-auth/cookies";
+import { ATLAS_MODES } from "@useatlas/types/auth";
 
 const authMode = process.env.NEXT_PUBLIC_ATLAS_AUTH_MODE ?? "";
+const VALID_MODES = new Set<string>(ATLAS_MODES);
 
 /** Routes that only unauthenticated users should see (exact match). */
 const authRoutes = ["/signup", "/login"];
@@ -32,17 +37,30 @@ function isAuthRoute(pathname: string): boolean {
   return authRoutes.some((route) => pathname === route);
 }
 
+/**
+ * Forward the `atlas-mode` cookie as an `x-atlas-mode` request header so
+ * server components can read the resolved mode without parsing cookies.
+ * Invalid or missing cookie values default to `published`.
+ */
+function withModeHeader(request: NextRequest): NextResponse {
+  const raw = request.cookies.get("atlas-mode")?.value;
+  const mode = raw && VALID_MODES.has(raw) ? raw : "published";
+  const headers = new Headers(request.headers);
+  headers.set("x-atlas-mode", mode);
+  return NextResponse.next({ request: { headers } });
+}
+
 export function proxy(request: NextRequest) {
   // Only enforce redirects for managed (Better Auth) mode
   if (authMode !== "managed") {
-    return NextResponse.next();
+    return withModeHeader(request);
   }
 
   const { pathname } = request.nextUrl;
 
   // Always allow public routes (demo, shared conversations, API, Next.js internals)
   if (isPublicRoute(pathname)) {
-    return NextResponse.next();
+    return withModeHeader(request);
   }
 
   try {
@@ -67,7 +85,7 @@ export function proxy(request: NextRequest) {
     );
   }
 
-  return NextResponse.next();
+  return withModeHeader(request);
 }
 
 // Next.js 16 requires the export to be named `config`, not `proxyConfig`.
